everything moved to database

+++ src/main/java/hu/joel/laczkovszki/qa/model/Hobby.java
@@ -3,19 +3,19 @@ package hu.joel.laczkovszki.qa.model;
 import java.util.List;
 
 public class Hobby {
-    private int id;
+    private Long id;
     private List<String> fieldsOfInterest;
 
-    public Hobby(int id, List<String> fieldsOfInterest) {
+    public Hobby(Long id, List<String> fieldsOfInterest) {
         this.id = id;
         this.fieldsOfInterest = fieldsOfInterest;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
